# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Account Documentation",
        "contact": {
            "name": "Marcos Azevedo",
            "email": "marcos.workspace@gmail.com"
        },
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "paths": {
...
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "operationId": "23679253d16f34d2972249751e3631dd",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 15
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UsersResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "operationId": "ce3cf31668a69eb39e5e98738715afa1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsersResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "operationId": "1ec60d4ee968175164a94c21f897f8ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsersResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "operationId": "786a794071cb7098f55a4a4911e7d3f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsersResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "operationId": "a56b12a3694a926699892056cdcf8a9e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UsersResource"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserStoreRequest": {
                "properties": {
                    "name": {
                        "format": "string"
                    },
                    "email": {
                        "format": "email"
                    },
                    "password": {
                        "format": "password"
                    },
                    "password_confirmation": {
                        "format": "password"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "UserUpdateRequest": {
                "properties": {
                    "name": {
                        "format": "string"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "PermissionsResource": {
                "description": "Class PermissionsResource",
                "properties": {
                    "id": {
                        "format": "int64"
                    },
                    "description": {
                        "format": "string"
                    },
                    "permission": {
                        "format": "string"
                    }
                },
                "type": "object"
            },
            "RolesResource": {
                "description": "Class RolesResource",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "description": {
                        "format": "string"
                    },
                    "createdAt": {
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "format": "date-time"
                    },
                    "permissions": {
                        "title": "Permissions",
                        "description": "permissions",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PermissionsResource"
                        }
                    }
                },
                "type": "object"
            },
            "UsersResource": {
                "description": "Class UsersResource",
                "properties": {
                    "id": {
                        "format": "int64"
                    },
                    "name": {
                        "format": "string"
                    },
                    "email": {
                        "format": "email"
                    },
                    "createdAt": {
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "format": "date-time"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RolesResource"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}